Fix paths that were broken due to folder rename.
